# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Platform API
  description: |
    社内ナレッジ活用プラットフォーム — BFF + マイクロサービス統合仕様書
    FR-01, FR-03, FR-04, FR-05, FR-06, FR-07, FR-08, FR-09, FR-10, FR-11, FR-13, FR-22,
    UC-01, UC-02, UC-05, UC-11
  version: "0.1.0"
  contact:
    name: Platform Team
servers:
  - url: http://localhost:5000
    description: BFF (ローカル開発)
  - url: http://localhost:5001
    description: DocumentService (直接)
  - url: http://localhost:5003
    description: RetrievalService (直接)
  - url: http://localhost:5004
    description: AiAnalysisService (直接)
  - url: http://localhost:5005
    description: AuthorizationService (直接)
  - url: http://localhost:5007
    description: "LlmGateway — IADR-0017: 内部 API。通常は host 非公開でサービス間呼び出し専用（compose では expose のみ）。この URL は個別に dotnet run した場合のみ到達可能"
...
paths:
...
  /bff/admin/authz/policies:
    get:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, SC-09: ABAC ポリシー一覧（BFF 集約）"
      operationId: bff-authz-list-policies
      responses:
        "200":
          description: ポリシー一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AbacPolicyDto"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "502": { description: AuthorizationService へ到達できない }
    post:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, UC-05, SC-09: ABAC ポリシー登録（保存前に矛盾検証）"
      description: |
        後段は保存前に属性辞書との整合を検証し、矛盾があれば **400（RFC7807 ValidationProblem。
        `errors.errors` にメッセージ配列）** で拒否する。BFF はこれを透過し、画面が検証結果として表示する。
      operationId: bff-authz-create-policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbacPolicyDto"
        "400":
          description: 保存前検証エラー（RFC7807 ValidationProblem）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationProblemDetails"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "502": { description: AuthorizationService へ到達できない }
  /bff/admin/authz/policies/validate:
    post:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-05, FR-09, SC-09（裁定 Q23 / #535）: ポリシーの dry-run 検証（保存しない）"
      description: |
        **保存せず検証だけ行う。** hi-fi が保存とは別に描く「検証」ボタンの後ろ側である
        （確定・2026-08-05・利用者裁定 Q23）。

        **保存経路（`POST /bff/admin/authz/policies`）と同じ検証を通る。**
        後段の 3 経路（dry-run / 登録 / 更新）が同一の検証関数を呼ぶことで一致を保証している。
        計画は「ローカルでの代用は採らない——『検証は通ったのに保存で矛盾が出る』形になり、
        検証ボタンへの信頼が失われる。**信頼できない検証ボタンは無いより悪い**」と定めている。

        **矛盾があっても 200 である**（`valid: false` ＋ `errors`）。検証した結果として矛盾が
        見つかったことは、要求の失敗ではない。**保存は従来どおり 400（RFC7807）** を返す。

        **要求型は登録と同じ `CreatePolicyRequest` である**——画面が保存用と検証用で 2 つの
        組み立てを持つと、そこがズレる余地になる。
      operationId: bff-authz-validate-policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "200":
          description: 検証結果（矛盾の有無によらず 200）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatePolicyResponse"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "502": { description: AuthorizationService へ到達できない }
  /bff/admin/authz/policies/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, SC-09: ABAC ポリシー個別取得（BFF 集約）"
      operationId: bff-authz-get-policy
      responses:
        "200":
          description: ポリシー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbacPolicyDto"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "404": { description: 見つからない }
        "502": { description: AuthorizationService へ到達できない }
    put:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, UC-05, SC-09: ABAC ポリシー更新（保存前に矛盾検証）"
      description: "後段は登録と同じ `CreatePolicyRequest` を受ける（専用の更新型を持たない）。"
      operationId: bff-authz-update-policy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbacPolicyDto"
        "400":
          description: 保存前検証エラー（RFC7807 ValidationProblem）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationProblemDetails"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "404": { description: 見つからない }
        "502": { description: AuthorizationService へ到達できない }
    delete:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, SC-09: ABAC ポリシー削除（BFF 集約）"
      operationId: bff-authz-delete-policy
      responses:
        "204": { description: 削除成功 }
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "404": { description: 見つからない }
        "502": { description: AuthorizationService へ到達できない }
  /bff/admin/authz/policies/{id}/active:
    patch:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, SC-09: ポリシー有効／無効の切替（削除せず一時停止）"
      operationId: bff-authz-set-policy-active
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetActiveRequest"
      responses:
        "200":
          description: 切替成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbacPolicyDto"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "404": { description: 見つからない }
        "502": { description: AuthorizationService へ到達できない }
  /bff/admin/authz/attributes:
    get:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, SC-09: 属性辞書一覧（BFF 集約）"
      operationId: bff-authz-list-attributes
      responses:
        "200":
          description: 属性辞書一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttributeDefinitionDto"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "502": { description: AuthorizationService へ到達できない }
    post:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, UC-05, SC-09: 属性辞書登録（キー重複・許可値検証）"
      operationId: bff-authz-create-attribute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttributeRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttributeDefinitionDto"
        "400":
          description: 保存前検証エラー（RFC7807 ValidationProblem。キー重複を含む）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationProblemDetails"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "502": { description: AuthorizationService へ到達できない }
  /bff/admin/authz/attributes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, SC-09: 属性辞書個別取得（BFF 集約）"
      operationId: bff-authz-get-attribute
      responses:
        "200":
          description: 属性辞書
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttributeDefinitionDto"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "404": { description: 見つからない }
        "502": { description: AuthorizationService へ到達できない }
    put:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, UC-05, SC-09: 属性辞書更新（Key / Scope は不変のため受け取らない）"
      operationId: bff-authz-update-attribute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAttributeRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttributeDefinitionDto"
        "400":
          description: 保存前検証エラー（RFC7807 ValidationProblem）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ValidationProblemDetails"
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "404": { description: 見つからない }
        "502": { description: AuthorizationService へ到達できない }
    delete:
      x-roles: [platform-admin]
      tags: [Authorization]
      summary: "FR-09, SC-09: 属性辞書削除（参照中は 409）"
      description: |
        FR-09, IADR-0006: 既存ポリシーが当該キーを参照している場合は削除を **409** で拒否する
        （辞書外チェックが効かなくなりポリシー条件の実効的制約が緩むのを防ぐ）。
        本文は RFC7807 の Problem で、**`detail` に参照元ポリシー名が列挙される**
        （「属性 '<key>' (scope=<scope>) は次のポリシーが参照しているため削除できません: <名前, …>」）。
        IADR-0040 決定 2: 画面はこの `detail` を拒否理由として表示する。
      operationId: bff-authz-delete-attribute
      responses:
        "204": { description: 削除成功 }
        "401": { description: 未認証 }
        "403": { description: 管理者ロール以外（AdminOnly） }
        "404": { description: 見つからない }
        "409":
          description: 参照中の属性は削除できない（`detail` に参照元ポリシー名）
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "502": { description: AuthorizationService へ到達できない }

  # ── SC-03/SC-05, FR-06, UC-01/UC-03/UC-07: 文書 BFF ─
  # IADR-0041 / IADR-0009: 読み取りはロールで塞がず、ABAC スコープ解決の結果でのみ絞る。
  # スコープ外・不在はいずれも 404 で応答し存在を秘匿する（「拒否」と「不在」を区別しない）。
  # 書き込みは **platform-admin のみ**（#629。グループ既定は「閲覧の下限」として
  # platform-admin / platform-operator を残し、各書き込み口へ AdminOnly を積んで AND 合成する）。
  # 対象がスコープ内であることを先に確かめる（閲覧できない文書は変更もできない）。
  # 後段の検証 400・競合 409・不在 404 は透過する。
...
  /authz/scope:
    post:
      tags: [Authorization]
      summary: "FR-05, ADR-0004: ABAC スコープ解決"
      operationId: resolve-scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccessScopeRequest"
      responses:
        "200":
          description: 許可されたフィルター
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessScopeResponse"
  /authz/policies:
    get:
      tags: [Authorization]
      summary: "FR-09: ABAC ポリシー一覧"
      operationId: list-policies
      responses:
        "200":
          description: ポリシー一覧
    post:
      tags: [Authorization]
      summary: "FR-09, UC-05: ABAC ポリシー登録（保存前に矛盾検証）"
      operationId: create-policy
      responses:
        "201":
          description: 作成成功
        "400":
          description: バリデーションエラー（RFC7807）
  /authz/policies/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Authorization]
      summary: "FR-09: ポリシー個別取得"
      operationId: get-policy
      responses:
        "200": { description: ポリシー }
        "404": { description: 見つからない }
    put:
      tags: [Authorization]
      summary: "FR-09, UC-05: ポリシー更新（保存前に矛盾検証）"
      operationId: update-policy
      responses:
        "200": { description: 更新成功 }
        "400": { description: バリデーションエラー（RFC7807） }
        "404": { description: 見つからない }
    delete:
      tags: [Authorization]
      summary: "FR-09: ポリシー削除"
      operationId: delete-policy
      responses:
        "204": { description: 削除成功 }
        "404": { description: 見つからない }
  /authz/policies/{id}/active:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Authorization]
      summary: "FR-09: ポリシー有効／無効の切替"
      operationId: set-policy-active
      responses:
        "200": { description: 切替成功 }
        "404": { description: 見つからない }
  /authz/attributes:
    get:
      tags: [Authorization]
      summary: "FR-09: 属性辞書一覧"
      operationId: list-attributes
      responses:
        "200": { description: 属性辞書一覧 }
    post:
      tags: [Authorization]
      summary: "FR-09, UC-05: 属性辞書登録（キー重複・許可値検証）"
      operationId: create-attribute
      responses:
        "201": { description: 作成成功 }
        "400": { description: バリデーションエラー（RFC7807） }
  /authz/attributes/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Authorization]
      summary: "FR-09: 属性辞書個別取得"
      operationId: get-attribute
      responses:
        "200": { description: 属性辞書 }
        "404": { description: 見つからない }
    put:
      tags: [Authorization]
      summary: "FR-09, UC-05: 属性辞書更新（Key/Scope は不変）"
      operationId: update-attribute
      responses:
        "200": { description: 更新成功 }
        "400": { description: バリデーションエラー（RFC7807） }
        "404": { description: 見つからない }
    delete:
      tags: [Authorization]
      summary: "FR-09: 属性辞書削除"
      operationId: delete-attribute
      responses:
        "204": { description: 削除成功 }
        "404": { description: 見つからない }
...
components:
  schemas:
...
    AccessScopeRequest:
      type: object
      required: [userId, userAttributes]
      properties:
        userId: { type: string }
        userAttributes:
          type: object
          additionalProperties: { type: string }
        action:
          type: string
          description: >-
            ［2026-08-23 / #989］解決したいアクション（`read` / `analyze` / `manage` / `write`。
            `enum` にしない——値域の正は `PolicyAction`）。省略時は `read`（旧クライアントは無改修で
            従来どおり読み取りスコープを得る）。従前はサーバ側が `read` をハードコードしており、
            書き込みの認可スコープをこの経路で出せなかった。値域外は 400（呼び出し側は非 2xx を
            deny へ縮退させる）。
    AccessScopeResponse:
      type: object
      description: |
        FR-05, ADR-0004: ABAC スコープ解決結果（`Platform.Shared.Contracts/Dtos/AccessScopeDto.cs`）。

        **`granted` と `allowedFilters` は組で読む。** 計画 `06_technical/07_abac-attribute-model.md`
        §ポリシー評価モデル の具体判定規則を、そのまま契約の 2 状態へ写したものである。

        | `granted` | `allowedFilters` | 意味 |
        | --- | --- | --- |
        | `false` | `[]` | **全件遮断**（利用者にマッチするポリシーが 1 件も無い＝deny-by-default） |
        | `true` | `[]` | **条件なしで全件許可**（マッチしたポリシーに文書条件が無い） |
        | `true` | 1 件以上 | 各フィルタを満たす文書だけ許可（フィルタ間は AND・値集合内は OR） |

        **`allowedFilters` が空であることだけでは 1 行目と 2 行目を区別できない。**
        `granted` を読まずに「フィルタが無い＝制限なし」と解釈すると **deny-by-default が全件開放へ反転する**。
        `granted` は既定値つきメンバー（`bool Granted = false`）だが**応答本文には必ず出る**ため
        `required` に含める（IADR-0132 論点 B の B1）。

        ［2026-08-10 / #525］`granted` はこの版で追加した。それ以前の本スキーマは
        `userId` と `allowedFilters` しか持たず、上表の 1 行目と 2 行目が同一の応答になっていた。

        ［2026-08-22 / #989］`branches` をこの版で追加した。**上表は `branches` が空／未指定のときの
        読み方であり、それは従来と変わらない。** 計画の `read` 規則は「静的属性ベース ∨ 所有者ベース ∨
        共有先ベース」の**選言**だが、`allowedFilters` は**単一の連言しか表せない**ため所有者ベースの
        判定が成立しなかった（`ADR-0046` D-06 部品 3）。`branches` はその選言を運ぶ。

        | `branches` | 読み方 |
        | --- | --- |
        | 空 / 未指定 | **従来どおり `allowedFilters` で評価する**（後方互換） |
        | 1 件以上 | **いずれかの分岐のフィルタをすべて満たす文書が可視**（分岐内 AND・分岐間 OR） |

        🔴 **`allowedFilters` は算出アルゴリズムごと据え置いてある。** 未移行の消費側は挙動が
        1 ビットも変わらない。**`allowedFilters`（分岐の積に相当）は `branches`（分岐の和）の
        部分集合であるため、未移行側が余分に見せることは構造上あり得ない** —— 移行中の乖離は
        常に deny 側へ倒れる。
      required: [userId, allowedFilters, granted]
      properties:
        userId: { type: string }
        granted:
          type: boolean
          description: >-
            許可ポリシーが 1 つでも一致したか。`false` は「閲覧可能な文書が無い」を意味し、
            `allowedFilters` が空でも**全件開放ではない**。
        allowedFilters:
          type: array
          items:
            type: object
            description: "FR-05: 1 つの属性キーに対する許可値集合（`AttributeFilter`）"
            required: [key, allowedValues]
            properties:
              key: { type: string }
              allowedValues:
                type: array
                items: { type: string }
        branches:
          type: array
          nullable: true
          description: >-
            FR-19, ADR-0036, ADR-0046 D-06: `read` の選言を運ぶ名前つき分岐（`AccessScopeBranch`）。
            **分岐内のフィルタは AND、分岐どうしは OR。** 空／未指定なら `allowedFilters` で評価する。
            `name` は監査・デバッグ用の識別子（`attribute` / `owner` / `shared`）であり、
            **どの分岐で可視になったかを言えるようにするために持たせている。**
          items:
            type: object
            required: [name, filters]
            properties:
              name: { type: string }
              filters:
                type: array
                items:
                  type: object
                  required: [key, allowedValues]
                  properties:
                    key: { type: string }
                    allowedValues:
                      type: array
                      items: { type: string }
...
    AbacPolicyDto:
      type: object
      description: "FR-09: ABAC ポリシー（利用者属性 × 文書属性 → 許可アクション）"
      required: [id, name, action, userConditions, documentConditions, isActive, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        action: { type: string, description: "`read` / `analyze` / `manage` / `write`（`enum` にしない。値域の正は `PolicyAction`）" }
        userConditions:
          $ref: "#/components/schemas/AbacConditionMap"
        documentConditions:
          $ref: "#/components/schemas/AbacConditionMap"
        isActive: { type: boolean, description: "削除せず一時停止できるようにするための有効フラグ" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AbacConditionMap:
      type: object
      description: |
        FR-09: 属性キー → 許可値の集合（例 `{"department": ["sales","hr"]}`）。
        契約が表現するのは**集合への所属**だけであり、比較演算子・包含などの式は表現しない。
        #520: **`required` を入れない**——`properties` を持たない写像（`additionalProperties` のみ）で、
        `required` は `properties` のキーを指すため適用できない。本スキーマは唯一「要求と応答の両方」で
        使われるが、この理由で要求側への波及も起こらない。
      additionalProperties:
        type: array
        items: { type: string }
    AttributeDefinitionDto:
      type: object
      description: "FR-09: 属性辞書エントリ（管理者が定義する取りうる値）"
      required: [id, key, label, allowedValues, required, scope, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        key: { type: string, example: "confidentiality" }
        label: { type: string }
        allowedValues:
          type: array
          items: { type: string }
        required: { type: boolean }
        scope: { type: string, description: "`document` / `user`。**Key と併せて一意**で、登録後は不変" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ValidatePolicyResponse:
      type: object
      description: |
        FR-05, FR-09, SC-09（裁定 Q23 / #535）: ポリシーの dry-run 検証の結果。
        **保存していない。** `valid: false` でも HTTP は 200 である
        （矛盾が見つかったことは要求の失敗ではない）。
        `errors` の内容は保存時の 400（`errors.errors`）と**同じ検証から出る**。
      required: [valid, errors]
      properties:
        valid: { type: boolean, description: 矛盾が無ければ true }
        errors:
          type: array
          items: { type: string }
          description: "矛盾の理由（日本語）。`valid: true` のときは空配列。"
    CreatePolicyRequest:
      type: object
      required: [name, action, userConditions, documentConditions]
      description: "FR-09, UC-05: ポリシーの登録・更新（更新も同じ型を用いる）"
      properties:
        name: { type: string }
        action: { type: string, description: "`read` / `analyze` / `manage` / `write`（`enum` にしない。値域の正は `PolicyAction`）" }
        userConditions:
          $ref: "#/components/schemas/AbacConditionMap"
        documentConditions:
          $ref: "#/components/schemas/AbacConditionMap"
    SetActiveRequest:
      type: object
      required: [isActive]
      properties:
        isActive: { type: boolean }
    CreateAttributeRequest:
      type: object
      required: [key, label, allowedValues, required]
      properties:
        key: { type: string }
        label: { type: string }
        allowedValues:
          type: array
          items: { type: string }
        required: { type: boolean }
        scope:
          type: string
          nullable: true
          description: "`document` / `user`。未指定は `document`"
    UpdateAttributeRequest:
      type: object
      required: [label, allowedValues, required]
      description: "FR-09: 属性辞書の更新。**Key / Scope は同一性・一意制約の基礎のため不変で、受け取らない**"
      properties:
        label: { type: string }
        allowedValues:
          type: array
          items: { type: string }
        required: { type: boolean }

    # ── RFC7807（Problem Details）──────────────────────
    ProblemDetails:
      type: object
      description: |
        RFC7807。`Results.Problem(title, detail, statusCode)` が返す形。
        #520: **`required` を入れない**——RFC7807 はどのメンバーも省略され得る（指定しなかった項目は
        出力されない）。ここで必須化すると、実在しない保証を生成型に持たせることになる。
      properties:
        type: { type: string, nullable: true }
        title: { type: string, nullable: true }
        status: { type: integer, nullable: true }
        detail: { type: string, nullable: true }
        instance: { type: string, nullable: true }
    ValidationProblemDetails:
      type: object
      description: |
        RFC7807 の検証エラー。`Results.ValidationProblem` が返す形で、
        AuthorizationService は `errors.errors` に**メッセージの配列**を束ねる
        （キーごとの配列という一般形に載せた運用。`apiClient.parseProblemDetails` がこれを平坦化する）。
        #520: **`required` を入れない**（理由は `ProblemDetails` と同じ）。
      properties:
        type: { type: string, nullable: true }
        title: { type: string, nullable: true }
        status: { type: integer, nullable: true }
        detail: { type: string, nullable: true }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }

    # ── FR-08: Feedback ───────────────────────────────
...